% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cube_probs_files}
\alias{cube_probs_files}
\title{Create a name for probs cube files}
\usage{
cube_probs_files(proj_dir, cube, version)
}
\arguments{
\item{proj_dir}{a \code{character} with project directory}

\item{cube}{a \code{sits_cube} object}

\item{version}{a \code{character} with version}
}
\value{
a \code{character} of cube probs files
}
\description{
Create a name for probs cube files
}
