% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{do_classify}
\alias{do_classify}
\title{}
\usage{
do_classify(
  proj_dir,
  cube,
  roi = NULL,
  version = "v1",
  memsize = 20,
  multicores = 20,
  force = FALSE
)
}
\arguments{
\item{proj_dir}{a \code{character} with project directory}

\item{cube}{a \code{sits_cube} object}

\item{roi}{a \code{list} with interested region}

\item{version}{a \code{character} with version}

\item{memsize}{a \code{numeric} with the memory size that will be used}

\item{multicores}{a \code{numeric} with number of cores will be used}

\item{force}{a \code{logical} if TRUE the samples will be overwritten}
}
\value{
an invisible NULL
}
\description{

}
