% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{do_get_ts}
\alias{do_get_ts}
\title{Get time series}
\usage{
do_get_ts(
  proj_dir,
  cube,
  csv_files,
  num_splits = 10,
  multicores = 20,
  force = FALSE
)
}
\arguments{
\item{proj_dir}{a \code{character} with project directory}

\item{cube}{a \code{sits_cube} object}

\item{csv_files}{a \code{character} with csv samples directory}

\item{num_splits}{a \code{numeric} with number of partitions will be splitted}

\item{multicores}{a \code{numeric} with number of cores will be used}

\item{force}{a \code{logical} if TRUE the samples will be overwritten}
}
\value{
an invisible NULL
}
\description{
Get time series
}
