% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{do_split}
\alias{do_split}
\title{Split samples}
\usage{
do_split(proj_dir, num_splits = 10)
}
\arguments{
\item{proj_dir}{a \code{character} with project directory}

\item{num_splits}{a \code{numeric} with number of partitions will be splitted}
}
\value{
an invisible NULL
}
\description{
Split samples
}
