% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{do_train}
\alias{do_train}
\title{Train model}
\usage{
do_train(proj_dir, ml_method, force = FALSE)
}
\arguments{
\item{proj_dir}{a \code{character} with project directory}

\item{ml_method}{a \code{sits_model} object}

\item{force}{a \code{logical} if TRUE the samples will be overwritten}
}
\value{
an invisible NULL
}
\description{
Train model
}
