% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extension}
\alias{extension}
\title{Get the extension name of file}
\usage{
extension(file)
}
\arguments{
\item{file}{a \code{character} with the name of file}
}
\value{
a \code{character} with the extension name of files
}
\description{
Get the extension name of file
}
