% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fix_cube_names_uniqueness}
\alias{fix_cube_names_uniqueness}
\title{Fix name of cubes}
\usage{
fix_cube_names_uniqueness(cube)
}
\arguments{
\item{cube}{a \code{sits_cube} object}
}
\value{
a \code{sits_cube} object with fixed name
}
\description{
Fix name of cubes
}
