% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_ts_from_csv}
\alias{get_ts_from_csv}
\title{Get time series from csv file}
\usage{
get_ts_from_csv(csv_file, cube, force = FALSE)
}
\arguments{
\item{csv_file}{\code{character} with csv samples directory}

\item{cube}{a \code{sits_cube} object}

\item{force}{a \code{logical} if TRUE the samples will be overwritten}
}
\value{
a \code{tibble} with time series extracted
}
\description{
Get time series from csv file
}
