% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_dir}
\alias{make_dir}
\title{Create directory}
\usage{
make_dir(...)
}
\arguments{
\item{...}{a  \code{character} vector with path}
}
\value{
a \code{character} vector with the path of the created directory
}
\description{
Create directory
}
