% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_proj_dir}
\alias{make_proj_dir}
\title{Creates project directory}
\usage{
make_proj_dir(proj_dir, samples_file, force = FALSE)
}
\arguments{
\item{proj_dir}{a \code{character} with project directory}

\item{samples_file}{a \code{character} with samples directory}

\item{force}{a \code{logical} if TRUE the samples will be overwritten}
}
\value{
an invisible NULL
}
\description{
Creates project directory
}
