% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{model_dir}
\alias{model_dir}
\title{Get the model directory}
\usage{
model_dir(proj_dir)
}
\arguments{
\item{proj_dir}{a \code{character} with project directory}
}
\value{
a \code{character} with path of model
}
\description{
Get the model directory
}
