% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{model_file_rds}
\alias{model_file_rds}
\title{Get the model directory in rds}
\usage{
model_file_rds(proj_dir)
}
\arguments{
\item{proj_dir}{a \code{character} with project directory}
}
\value{
a \code{character} with path of model
}
\description{
Get the model directory in rds
}
