% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{probs_dir}
\alias{probs_dir}
\title{Get the probability images directory}
\usage{
probs_dir(proj_dir)
}
\arguments{
\item{proj_dir}{a \code{character} with project directory}
}
\value{
a \code{character} with path of probability images
}
\description{
Get the probability images directory
}
