% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{probs_file_rds}
\alias{probs_file_rds}
\title{Get the probability images directory in rds}
\usage{
probs_file_rds(proj_dir)
}
\arguments{
\item{proj_dir}{a \code{character} with project directory}
}
\value{
a \code{character} with path of probability images
}
\description{
Get the probability images directory in rds
}
