% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{samples_file_rds}
\alias{samples_file_rds}
\title{Get the path of samples in rds}
\usage{
samples_file_rds(proj_dir)
}
\arguments{
\item{proj_dir}{a \code{character} with project directory}
}
\value{
a \code{character} with path of samples
}
\description{
Get the path of samples in rds
}
