% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitsdraft.R
\name{sitsdraft}
\alias{sitsdraft}
\title{Creates a template package based on reproducible research to put sits
 analysis.}
\usage{
sitsdraft(
  pkgname = "sitsdraft",
  template = "template/classification.R",
  filename = "classification.R",
  rstudio = TRUE,
  open = TRUE,
  simple = FALSE
)
}
\arguments{
\item{pkgname}{a \code{character} with the path and package name}

\item{template}{a \code{character} with the path of the template file}

\item{filename}{a \code{character} with the name of template file}

\item{rstudio}{verify if the user uses rstudio, if so, a new project
file will be created.}

\item{open}{a \code{boolean} to open a new session after creating the
template. By default is TRUE.}

\item{simple}{a \code{boolean} to create an R directory in template
package. if FALSE the R directory will be created. By default is FALSE.}
}
\value{
After creating the template, if the user are in rstudio, the session
will be redirect to the created template.
}
\description{
This functions uses as basis the package \code{rrtools} to
 create a template package that follows a base structure of reproducible
 researcher package.
}
