% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{splits_dir}
\alias{splits_dir}
\title{Get the path of samples that will be spitted}
\usage{
splits_dir(proj_dir)
}
\arguments{
\item{proj_dir}{a \code{character} with project directory}
}
\value{
a \code{character} with path of samples
}
\description{
Get the path of samples that will be spitted
}
